+++ src/ducks/Studio/Widget/TopTransactionsTable/index.module.scss
